BackEndApi/requirements/base.txt
# Base
pytz
python-slugify
jinja2

# Django
Django==3.2
django-cors-headers==3.7.0
django-rest-passwordreset

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.4
drf-nested-routers==0.93.4
django-filter==2.4.0

# Image generator
qrcode==7.4.2

djangorestframework-simplejwt==5.2.2
pyjwt==1.7.1

# Passwords security
argon2-cffi==21.3.0

# Static files
whitenoise==5.2.0

# AWS
boto3==1.17.64

# Storage
psycopg2-binary==2.9.3
django-storages

# Channels
channels
channels_redis

# Web3
web3==5.31.3

# CMS
django-modeltranslation

# Images
pillow==9.4.0
requests

# Currencies
django-money

# interface
django-admin-interface

#sendgrip
django-sendgrid-v5==1.2.2

# OpenAI
openai
pydantic==2.4.2
httpx<0.28